+++ src/pages/feed/ui/index.ts
@@ -7,3 +7,4 @@ export { default as CdViewerLayout } from './layout/CdViewerLayout'
 export { default as FollowLayout } from './layout/FollowLayout'
 export { default as FeedCdListLayout } from './layout/FeedCdListLayout'
 export { default as FollowEmpty } from './FollowEmpty'
+export { default as FollowList } from './FollowList'
